test(component): move renderWithI18n into tests/fixtures, comment two attribute-only assertions
`tests/helpers/renderWithI18n.tsx` wasn't one of the two locations
CLAUDE.md names for shared test infrastructure. Moved it into
tests/fixtures and updated all 18 component-test imports.

Also added the rationale comment CLAUDE.md's "assert on behaviour, not
on how it happens" section requires for two attribute-only assertions
that had none: settings.spec.ts's data-theme check (ties it to the CSS
custom properties src/index.css keys on it) and JsonImport.test.tsx's
short-field maxlength check (notes that, unlike the paste box, nothing
else in the form validates that field's size).
