+++ tests/component/Avatar.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { Avatar } from '@/components/Avatar';
-import { renderWithI18n as render } from '../helpers/renderWithI18n';
+import { renderWithI18n as render } from '../fixtures/renderWithI18n';
 
 describe('Avatar', () => {
   it('shows the picture the identity provider supplied', () => {
