+++ tests/component/SpeakButton.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, screen } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 import { SpeakButton, SpeechStatusNote } from '@/components/SpeakButton';
-import { renderWithI18n as render } from '../helpers/renderWithI18n';
+import { renderWithI18n as render } from '../fixtures/renderWithI18n';
 
 /**
  * The pronunciation control on the word detail screens, where — unlike the
